requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
email-validator==2.1.0.post1
SQLAlchemy==2.0.27
Werkzeug==3.0.1
Jinja2==3.1.3
click==8.1.8
itsdangerous==2.2.0
MarkupSafe==2.1.5
psycopg2-binary==2.9.10
gunicorn==21.2.0
asgiref==3.8.1
chardet==5.2.0
crispy-bootstrap5==2025.4
Django==5.2
django-crispy-forms==2.4
django-tailwind==4.0.1
pillow==11.2.1
reportlab==4.3.1
sqlparse==0.5.3
tzdata==2025.2
cssselect2>=0.7.0
pydyf>=0.8.0
tinycss2>=1.2.0
fonttools>=4.38.0
requests >=2.32.3
django-cors-headers >= 4.7.0
djangorestframework >= 3.15
flask-cors >= 5.0.1
flask_migrate